Struct pqbus::Queue

pub struct Queue<'a, B> { /* fields omitted */ }

A named message queue

Methods

impl<'a, B> Queue<'a, B>

A push pop message queue.

fn size(&self) -> BusResult<i64>

Returns the number of messages in the queue.

fn is_empty(&self) -> BusResult<bool>

Determines if there are any pending messages.

fn push<E>(&self, obj: B) -> Result<(), PushError<E>> where B: ToMessageBody<E>

Pushes a message into the queue.

fn pop_blocking<E>(&self) -> Result<B, PopError<E>> where B: FromMessageBody<E>

Pops a message from the queue. Blocks if there are none pending.

fn pop_wait<E>(&self, timeout: Duration) -> Result<Option<B>, PopError<E>> where B: FromMessageBody<E>

Pops a message from the queue. Blocks for duration of timeout if there are none pending.

fn pop_callback<F, E>(&self, work_fn: F) -> Result<bool, BusError> where F: Fn(B), B: FromMessageBody<E>, E: Display

Run a closure on messages in the queue. Blocks if there are none pending.

fn pop<E>(&self) -> Result<Option<B>, PopError<E>> where B: FromMessageBody<E>

Pops a message from the queue if there is one pending.

fn messages<'queue, E>(&'queue self)
-> MessageIter<'a, 'queue, NextMessagePending, B, E> where B: FromMessageBody<E>

Returns an iterator over pending messages. Ends when the queue is empty.

fn messages_blocking<'queue, E>(&'queue self)
-> MessageIter<'a, 'queue, NextMessageBlocking, B, E> where B: FromMessageBody<E>

Returns an iterator over messages that blocks until a message is received if none are pending. This function never returns.